\name{RandomFactorTensor-package}
\alias{RandomFactorTensor-package}
\alias{RandomFactorTensor}
\docType{package}
\title{
\packageTitle{RandomFactorTensor}
}
\description{
\packageDescription{RandomFactorTensor}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{RandomFactorTensor}
\packageIndices{RandomFactorTensor}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{RandomFactorTensor}

Maintainer: \packageMaintainer{RandomFactorTensor}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
